Sumitting luigi tasks via subprocess
Because in "production" we are using uwsig to run python using
sys.executable will not work. In an uwsig environment sys.executable
return the uwsgi executable. Because of that the virtualenv home
is retrieved and the python executable is found. Some cleanup was
done. Removed the cli because I honestly don't know if it is
still running.